requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.9.0
blis==1.3.3
cachetools==6.2.6
catalogue==2.0.10
certifi==2026.1.4
cffi==1.15.1
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
comm @ file:///home/task_176311912485839/conda-bld/comm_1763119193329/work
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.13
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoints==0.4
exceptiongroup==1.2.2
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
iniconfig==2.0.0
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.1.8
jedi==0.18.1
Jinja2==3.1.2
joblib==1.5.3
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab_widgets==3.0.16
kagglesdk==0.1.15
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.1.6
mistune==2.0.4
murmurhash==1.0.15
narwhals==2.16.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
opencv-python==4.6.0.66
packaging==26.0
pandas==2.3.3
pandas-stubs==2.3.3.260113
pandocfilters==1.5.0
parso==0.8.3
pexpect @ file:///home/task_176253513056498/conda-bld/pexpect_1762535937673/work
pickleshare==0.7.5
pillow==12.1.0
platformdirs @ file:///home/task_176235608007063/conda-bld/platformdirs_1762356488148/work
pluggy==1.5.0
preshed==3.0.12
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==6.33.5
psutil==5.9.3
ptyprocess @ file:///opt/miniconda3/conda-bld/ptyprocess_1762424170819/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=3470be7f810474c8a2ecfcd6e02acc6aea8483ab595417fa4e336362a349933e
pure-eval==0.2.2
pyarrow==23.0.0
pycparser==2.21
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.13.0
pyparsing==3.3.2
pyrsistent==0.19.2
pytest==8.3.3
python-dateutil==2.8.2
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.2.1
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
Send2Trash==1.8.0
six @ file:///croot/six_1744271502820/work
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stack-data==0.6.0
streamlit==1.53.1
tenacity==9.1.2
terminado==0.17.0
thinc==8.3.10
threadpoolctl==3.6.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.2
tornado==6.2
tqdm==4.67.2
traitlets==5.5.0
typer-slim==0.21.1
types-pytz==2025.2.0.20251108
typing-inspection==0.4.2
typing_extensions @ file:///croot/typing_extensions_1756280817316/work
tzdata==2025.3
urllib3==2.6.3
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.5
weasel==0.4.3
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.15
wrapt==2.0.1
